- Test setup update for visibility check
