--- rs/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod test_document;
-pub(crate) mod test_power_point;
