+++ rs_test/src/main.rs
@@ -1,4 +1,6 @@
-mod tests;
+mod document;
+mod presentation;
+mod spreadsheet;
 
 fn main() {
     println!("Hello, world!");
