--- rs_test/src/tests/mod.rs
@@ -1,1 +1,0 @@
-mod excel;
